requirements.txt
PyNaCl
discord
youtube_dl
python-dotenv
